requirements.txt
requests>=2.28.2
pillow>=9.5.0
piexif>=1.1.3
pytz>=2023.3
pyyaml>=6.0
mutagen>=1.46.0
tqdm>=4.65.0